FEATURE: PROVISION_LOCATION was changed to Location.java
Currently SoftwareProcess allows to run the entity in a MachineLocation. These changes add to
SoftwareProcess supporting for any kind of location, for example PaasLocation.

Then, the attribute PROVISION_LOCATION was rename as LOCATION the the type was changed to Location.java.

SoftwareProcessImpl contains a set of effectors attached from MachineLifecycleEffectorTasks, these effectors
are developed for a MachineLocation (as the MachineLifecycleEffectorTasks' name indicates), so into this class
the attribute SoftwareProcess.LOCATION is casted as (MachineProvisioningLocation).