Extracting ssh behaviour from SoftwareProcessImpl: provisioning flags
For adding paas support, SoftwareProcess entities should to be agnostics in order to allows to be deployed
on SSH and PaaS locations. So, the machine behaviour and management should be decoupled of the entities
and it should be moved to an isolate class.

SoftwareProcessImpl contained a pair of method for obtaining the needed flags to create and configure a VM.
However as it was mentioned previously, this entity shouldn't contain any ssh (machine) behaviour. Then, this
code is moved to an external class, MachineProvisioningLocationFlagsSupplier that is managed in
SoftwareProcessImpl as a LocationFlagSupplier interface. Moreover, an equivalent class is added to represent
this behaviour (obtain flags location) for PaaS.

SoftwareProcessImpl will select a behaviour for obtaining the flags depending on the selected location for deploying
the entity.