Adding PaaS support to TomcatServer entity.
The drivers for CloudFoundry are added to SoftwareProcess entity and JavaWebAppSoftwareProcess implementation entity.
The final PaaS support is added to TomcatServer entity using a new driver which is focused on CloudFoundryPaaSLocation.

A new hierarchy of drivers is added to Brooklyn, these new drivers is oriented to PaaS but they follow the hierarchy
defined for SSH. For example, AbstractSoftwareProcessDriver, AbstractSoftwareProcessSshDriver allow to SoftwareProcess entities
use drivers for location based on machines. Then, to enable the PaaS support for SoftwareProcess entities a set of drivers is added such as
AbstractSoftwareProcessCloudFoundryDriver and AbstractApplicationCloudFoundryDriver.

Several features in different entities are modified to allows the PaasLocation management.